back/requirements.txt
asgiref==3.8.1
bidict==0.23.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
eventlet==0.38.0
greenlet==3.1.1
h11==0.14.0
psycopg2==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
sqlparse==0.5.2
tzdata==2024.2
wsproto==1.2.0
